Move files to session
