+++ onnxruntime/core/session/plugin_ep/ep_api.cc
@@ -7,7 +7,6 @@
 #include <vector>
 
 #include "core/common/semver.h"
-#include "core/framework/abi_kernel.h"
 #include "core/framework/error_code_helper.h"
 #include "core/framework/func_api.h"
 #include "core/framework/ort_value.h"
@@ -20,6 +19,7 @@
 #include "core/session/abi_ep_types.h"
 #include "core/session/onnxruntime_ep_device_ep_metadata_keys.h"
 #include "core/session/ort_apis.h"
+#include "core/session/plugin_ep/ep_kernel_registration.h"
 
 using namespace onnxruntime;
 namespace OrtExecutionProviderApi {
