+++ onnxruntime/core/session/plugin_ep/ep_plugin_provider_interfaces.cc
@@ -9,7 +9,6 @@
 #include <unordered_set>
 #include <utility>
 #include <vector>
-#include "core/framework/abi_kernel.h"
 #include "core/framework/compute_capability.h"
 #include "core/framework/error_code_helper.h"
 #include "core/framework/model_metadef_id_generator.h"
@@ -22,7 +21,7 @@
 #include "core/session/abi_logger.h"
 #include "core/session/abi_session_options_impl.h"
 #include "core/session/allocator_adapters.h"
-#include "core/session/plugin_ep/ep_api.h"
+#include "core/session/plugin_ep/ep_kernel_registration.h"
 #include "core/session/ort_apis.h"
 #include "core/providers/partitioning_utils.h"
 
@@ -126,52 +125,6 @@ static const Node* FindFirstNodeAssignedToOtherEP(const std::string& ep_type,
                                 });
 
   return node_iter != ep_nodes.end() ? &(*node_iter)->GetInternalNode() : nullptr;
-}
-
-static Status InitKernelRegistry(OrtEp& ort_ep, /*out*/ std::shared_ptr<KernelRegistry>& kernel_registry) {
-  kernel_registry = nullptr;
-
-  if (ort_ep.GetNumKernelCreateInfos == nullptr || ort_ep.GetKernelCreateInfos == nullptr) {
-    return Status::OK();
-  }
-
-  size_t num_kernels = ort_ep.GetNumKernelCreateInfos(&ort_ep);
-  if (num_kernels == 0) {
-    return Status::OK();
-  }
-
-  std::vector<OrtKernelCreateInfo*> kernel_create_infos(num_kernels);
-  Status status = ToStatusAndRelease(ort_ep.GetKernelCreateInfos(&ort_ep, kernel_create_infos.data(), num_kernels));
-
-  // Store OrtKernelCreateInfo provided by the plugin EP in std::unique_ptr so that they are always properly released.
-  std::vector<std::unique_ptr<OrtKernelCreateInfo, decltype(&OrtExecutionProviderApi::ReleaseKernelCreateInfo)>>
-      kernel_create_infos_holder;
-
-  kernel_create_infos_holder.reserve(num_kernels);
-  for (OrtKernelCreateInfo* kernel_create_info : kernel_create_infos) {
-    auto holder = std::unique_ptr<OrtKernelCreateInfo, decltype(&OrtExecutionProviderApi::ReleaseKernelCreateInfo)>(
-        kernel_create_info, OrtExecutionProviderApi::ReleaseKernelCreateInfo);
-    kernel_create_infos_holder.push_back(std::move(holder));
-  }
-
-  if (!status.IsOK()) {
-    return status;
-  }
-
-  // Add all KernelCreateInfo instances to the KernelRegistry
-  kernel_registry = std::make_shared<KernelRegistry>();
-
-  for (size_t i = 0; i < num_kernels; i++) {
-    OrtKernelCreateInfo* ort_kernel_create_info = kernel_create_infos[i];
-    PluginEpKernelCreateFunctor kernel_create_functor(ort_kernel_create_info->kernel_create_func,
-                                                      ort_kernel_create_info->kernel_create_func_state);
-    KernelCreateInfo kernel_create_info(std::make_unique<KernelDef>(ort_kernel_create_info->kernel_def),  // copy
-                                        kernel_create_functor);
-
-    ORT_RETURN_IF_ERROR(kernel_registry->Register(std::move(kernel_create_info)));
-  }
-
-  return Status::OK();
 }
 
 /*static*/
